# api/openapi.yaml
openapi: 3.1.0
info:
  title: GymBuddy API
  version: 1.0.0
  description: |
    GymBuddy is offline-first: every client holds a full local replica of
    the user's data and queues writes in a local outbox. The UI reads and
    writes that replica directly and never awaits a network call - sync
    with the server happens separately, on its own schedule, as
    push-then-pull (never concurrent).

    This is why there are no per-entity CRUD endpoints. The outbox is the
    only write path, for every client, online or offline; a second "live"
    write endpoint would give the server two places to enforce the same
    change-log and revision invariants, which is exactly the kind of
    drift that produces unreconcilable data. Writes go through
    `/v1/sync/push`, always. Conflict resolution is server-authoritative:
    the server decides who wins and stamps `updatedAt` and `revision`.
    On a win, push's own response carries those values back and the
    client overwrites its local copy with them directly. On a loss
    (`CONFLICT`), push returns no entity data at all - there is nothing
    to overwrite with yet. The client discards its local change and
    picks up the winning version from the next `/v1/sync/pull`, the only
    path a full entity ever travels back to a client. Clients never
    merge. Workout sessions and workout templates sync as whole
    aggregates together with their children (performed exercises, sets,
...
    it.

...

...
    the wire.
  contact:
...
servers:
...
security:
  - firebaseAuth: [ ]
...
paths:
  /v1/sync/push:
    post:
      tags: [ Sync ]
      summary: Push local changes
      description: |
        Sends locally-created, locally-modified, and locally-deleted
        entities. Each array is applied independently; ordering between
        arrays is not significant.

        A deletion is not a separate operation: it is an ordinary entity
        push with deletedAt set to any non-null value. The server treats
        this exactly like any other field change - same revision check,
        same idempotency rule - and stamps its own deletedAt on write;
        the value the client sends is only a signal, never stored
        verbatim. This does not apply to userProfile (account deletion
        is a separate, dedicated endpoint) or to userExerciseStates/
        userTemplateStates, which have no deletedAt at all.

        Every entity is applied in its own transaction and receives its own
        result, so one rejected entity does not block the rest of the batch.

        Idempotent by entity id. Re-pushing an entity the server has already
        applied returns the existing result rather than duplicating it.

        Inspect every result before clearing the outbox entry it came from:

          APPLIED   overwrite the local updatedAt and revision with the
                    server-assigned values returned here, then clear the
                    outbox entry.
          CONFLICT  the server held a newer version. This result carries
                    no entity data - updatedAt and revision are both
                    null - so there is nothing here to overwrite local
                    state with. Discard the local change, clear the
                    outbox entry, and wait for the next pull to deliver
                    the winning version; do not retry this push.
          INVALID   the entity fails validation and can never succeed as
                    sent - clear the outbox entry without retrying.
          FORBIDDEN the entity is not owned by the caller - clear the
                    outbox entry without retrying.
          ERROR     an unexpected server-side failure while processing
                    this entity - not the client's fault and not a
                    verdict on the data. Leave the outbox entry in place
                    and retry the push later; do not discard it.
      operationId: pushChanges
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushRequest'
            example:
              workoutSessions:
                - id: 018f5e9a-7c3d-7abc-8b2e-1a2b3c4d5e6f
                  startedAt: 1732628400000
                  endedAt: 1732631580000
                  durationSeconds: 3180
                  title: Push Day
                  notes: Felt strong today
                  templateId: 019fe506-35d6-75b5-a3cf-3bd2184cad28
                  templateTitle: Chest & Triceps
                  performedExercises:
                    - id: 018f5e9a-8d4e-7bcd-9c3f-2b3c4d5e6f70
                      exerciseId: 019fe506-359a-7601-92f2-f441cf1dae33
                      orderIndex: 0
                      exerciseName: Barbell Bench Press
                      exerciseCategory: STRENGTH
                      exerciseTrackingType: WEIGHT_REPS
                      exercisePrimaryMuscles: [ CHEST ]
                      sets:
                        - id: 018f5e9a-8e10-7011-8f21-5e6f70819203
                          weightKg: 60
                          reps: 8
                          setType: WARMUP
                          isCompleted: true
                          restTakenSeconds: 60
                          orderIndex: 0
                        - id: 018f5e9a-8e10-7222-9032-6f7081920314
                          weightKg: 80
                          reps: 6
                          setType: WORKING
                          isCompleted: true
                          restTakenSeconds: 120
                          orderIndex: 1
                        - id: 018f5e9a-8e10-7433-a143-708192031425
                          weightKg: 80
                          reps: 5
                          setType: WORKING
                          isCompleted: true
                          restTakenSeconds: 120
                          orderIndex: 2
                    - id: 018f5e9a-9e5f-7cde-ad40-3c4d5e6f7081
                      exerciseId: 019fe506-359b-7a6f-82f9-90f196c28c0a
                      orderIndex: 1
                      exerciseName: Barbell Back Squat
                      exerciseCategory: STRENGTH
                      exerciseTrackingType: WEIGHT_REPS
                      exercisePrimaryMuscles: [ LEGS ]
                      sets:
                        - id: 018f5e9a-9f60-7544-b254-819203142536
                          weightKg: 100
                          reps: 5
                          setType: WORKING
                          isCompleted: true
                          restTakenSeconds: 150
                          orderIndex: 0
                        - id: 018f5e9a-9f60-7655-c365-920314253647
                          weightKg: 100
                          reps: 5
                          setType: WORKING
                          isCompleted: true
                          restTakenSeconds: 150
                          orderIndex: 1
              exercises:
                - id: 018f5e9a-a16a-7def-be51-4d5e6f708192
                  name: Cable Lateral Raise (Behind Back)
                  primaryMuscles: [ SHOULDERS ]
                  secondaryMuscles: [ ]
                  description: A cable variation of the lateral raise, routed behind the back for constant tension through the full range of motion.
                  instructions:
                    - Stand sideways to a low cable pulley
                    - Grip the handle with the far hand, arm across the body
                    - Raise the arm out to the side to shoulder height
                    - Lower with control
                  tips:
                    - Keep a slight bend in the elbow
                    - Lead with the elbow, not the hand
                  difficulty: INTERMEDIATE
                  equipmentNeeded: [ CABLE ]
                  category: STRENGTH
                  exerciseType: ISOLATION
                  trackingType: WEIGHT_REPS
                  videoUrl: null
                  thumbnailUrl: null
                  source: CUSTOM
                  ownerId: 8Kj3mNpQR5X9YbZ2LtWvAeHc3Fg1
                  derivedFromId: null
                  deprecated: false
      responses:
        '200':
          description: Batch processed. Inspect per-entity results.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushResponse'
              example:
                results:
                  - entityType: WORKOUT_SESSION
                    entityId: 018f5e9a-7c3d-7abc-8b2e-1a2b3c4d5e6f
                    status: APPLIED
                    updatedAt: 1732631605123
                    revision: 1
                  - entityType: EXERCISE
                    entityId: 018f5e9a-a16a-7def-be51-4d5e6f708192
                    status: CONFLICT
                    reason: Server has a newer revision of this exercise, pushed from another device.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/sync/pull:
    get:
      tags: [ Sync ]
      summary: Pull changes since a cursor
      description: |
        Returns entities changed since the given cursor. Omit the cursor for
        a full initial sync.

        Pulling is a loop, not a single call: commit the page to local
        storage, persist nextCursor, then call again with that cursor if
        hasMore is true. Repeat until hasMore is false.

        The order matters. nextCursor must be persisted only after the
        page it came with has been committed locally - never before, and
        never together in a way that could commit one without the other.
        If the app is killed between persisting the cursor and committing
        the page, the next pull resumes from a cursor that already thinks
        this page was seen, and the data in it is silently lost. Persisting
        after commit means a crash just repeats the page - safe, because
        push/pull application is idempotent.

        A 410 means the cursor predates the tombstone retention window
        (deletions are only kept for 90 days) and can no longer be resumed
        from. The client must discard all local data for this user and
        start over by calling pull again with no cursor.

        This is evaluated against what this account actually still has on
        record, not a blanket age limit on the cursor value itself: a
        cursor that has nothing new to deliver never expires merely from
        going unused, no matter how long it sits - only cursors that
        would need to replay a deletion that has since been pruned are
        rejected. An account that rarely syncs is not penalised for it.

        A 400 means the cursor string itself couldn't be decoded - it
        wasn't produced by a previous pull, or was corrupted in storage.
        This is not the same as 410: a malformed cursor's position is
        unknown, not just old, so there's nothing to compare it against.
        Treat it the same way as a 410 in practice - discard local data
        and start over - but the two are logged and reported separately
        since they indicate different failures (client-side corruption
        vs. server-side retention).
      operationId: pullChanges
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: |
            Opaque token from a previous pull. Clients must treat this as
            an opaque string: never parse it, construct it, or compare two
            cursors for ordering. Omit for a full initial sync.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 200
            maximum: 500
      responses:
        '200':
          description: Changes since the cursor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullResponse'
              example:
                exercises:
                  - id: 018f5e9a-a16a-7def-be51-4d5e6f708192
                    name: Cable Lateral Raise (Behind Back)
                    primaryMuscles: [ SHOULDERS ]
                    secondaryMuscles: [ ]
                    description: A cable variation of the lateral raise, routed behind the back for constant tension through the full range of motion.
                    instructions:
                      - Stand sideways to a low cable pulley
                      - Grip the handle with the far hand, arm across the body
                      - Raise the arm out to the side to shoulder height
                      - Lower with control
                    tips:
                      - Keep a slight bend in the elbow
                      - Lead with the elbow, not the hand
                    difficulty: INTERMEDIATE
                    equipmentNeeded: [ CABLE ]
                    category: STRENGTH
                    exerciseType: ISOLATION
                    trackingType: WEIGHT_REPS
                    videoUrl: null
                    thumbnailUrl: null
                    source: CUSTOM
                    ownerId: 8Kj3mNpQR5X9YbZ2LtWvAeHc3Fg1
                    derivedFromId: null
                    deprecated: false
                    updatedAt: 1732631605123
                    deletedAt: null
                    revision: 1
                workoutTemplates:
                  - id: 019fe506-4210-7a3b-9c44-1a2b3c4d5e6f
                    title: Old Push Day
                    templateExercises: [ ]
                    source: CUSTOM
                    ownerId: 8Kj3mNpQR5X9YbZ2LtWvAeHc3Fg1
                    derivedFromId: null
                    deprecated: false
                    updatedAt: 1732631700000
                    deletedAt: 1732631700000
                    revision: 3
                nextCursor: eyJ2IjoxLCJtb2RlIjoiREVMVEEiLCJzZXEiOjI0NX0
                hasMore: true
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '410': { $ref: '#/components/responses/CursorExpired' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    firebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token. Verified server-side against Google's JWKS.
  schemas:
    PushRequest:
      type: object
      description: |
        Each array carries entities of one type. All are optional; omit or
        send empty for types with no pending changes.
      properties:
        workoutSessions:
          type: array
          items: { $ref: '#/components/schemas/WorkoutSession' }
        exercises:
          type: array
          description: Custom exercises only. Default exercises are server-owned.
          items: { $ref: '#/components/schemas/Exercise' }
        workoutTemplates:
          type: array
          items: { $ref: '#/components/schemas/WorkoutTemplate' }
        userExerciseStates:
          type: array
          items: { $ref: '#/components/schemas/UserExerciseState' }
        userTemplateStates:
          type: array
          items: { $ref: '#/components/schemas/UserTemplateState' }
        userProfile:
          $ref: '#/components/schemas/UserProfile'
    PushResponse:
      type: object
      required: [ results ]
      properties:
        results:
          type: array
          items: { $ref: '#/components/schemas/MutationResult' }
...
    PullResponse:
      type: object
      required: [ nextCursor, hasMore ]
      description: |
        Changed entities since the cursor. Deletions arrive as entities with
        a non-null deletedAt, not as a separate list - one code path handles
        both a create/update and a delete.

        limit bounds how many change_log rows the server scans per page,
        not how many entities come back. An entity changed several times
        since the cursor collapses to one entry here, but still counts as
        several against limit - so a page can legitimately contain fewer
        entities than limit, or even zero, while hasMore is still true.
        Never infer "nothing left" from a short or empty page; only
        hasMore says that.
      properties:
        workoutSessions:
          type: array
          items: { $ref: '#/components/schemas/WorkoutSession' }
        exercises:
          type: array
          items: { $ref: '#/components/schemas/Exercise' }
        workoutTemplates:
          type: array
          items: { $ref: '#/components/schemas/WorkoutTemplate' }
        userExerciseStates:
          type: array
          items: { $ref: '#/components/schemas/UserExerciseState' }
        userTemplateStates:
          type: array
          items: { $ref: '#/components/schemas/UserTemplateState' }
        userProfile:
          $ref: '#/components/schemas/UserProfile'
        nextCursor:
          type: string
          description: Persist only after committing this page locally.
        hasMore:
          type: boolean
...
    WorkoutSession:
      type: object
      description: |
        A completed or in-progress workout. Syncs as a whole aggregate
        together with its performed exercises and sets - there is no
        separate sync path for a child row.
      required:
        - id
        - startedAt
        - durationSeconds
        - title
        - performedExercises
        - updatedAt
        - revision
      properties:
        id:
          type: string
          format: uuid
          description: Client-generated UUIDv7. Doubles as the sync idempotency key.
        startedAt:
          type: integer
          format: int64
          description: Epoch milliseconds when the session started.
        endedAt:
          type: [integer, "null"]
          format: int64
          description: Epoch milliseconds when the session ended. Null means the session is still in progress.
        durationSeconds:
          type: integer
          format: int32
          description: |
            Total duration of the session, in seconds. Deliberately not
            derived from endedAt - startedAt: elapsed wall-clock time
            includes rest/idle time and the user may edit the duration
            independently of when the session actually started or ended.
        title:
          type: string
          description: Session title, e.g. a template name or "Workout".
        notes:
          type: [string, "null"]
          description: Per-session commentary, e.g. "felt strong today" or "shoulder was tight".
        templateId:
          type: [string, "null"]
          format: uuid
          description: |
            Soft reference to WorkoutTemplate.id this session was started
            from - not a foreign key, must tolerate the referenced template
            being deleted. Provenance only; use templateTitle for display.
        templateTitle:
          type: [string, "null"]
          description: |
            Snapshot of WorkoutTemplate.title at the time this session was
            started. Written once and never updated, so a later rename or
            deletion of the template does not rewrite history. Null if this
            session wasn't started from a template.
        performedExercises:
          type: array
          description: |
            Exercises performed during this session. Order reflects the
            sequence in which they were performed.
          items: { $ref: '#/components/schemas/PerformedExercise' }
        updatedAt:
          type: integer
          format: int64
          readOnly: true
          description: Server-assigned last-change time, epoch milliseconds. Clients must not set this.
        deletedAt:
          type: [integer, "null"]
          format: int64
          description: |
            Tombstone time, epoch milliseconds; non-null means this
            session is deleted. On push, this is the client's delete
            signal, not a timestamp the server trusts: set it to any
            non-null value to delete the session, leave it null
            otherwise. The server always stamps its own current time on
            write and ignores the value sent - what comes back from pull
            is the authoritative deletedAt.
        revision:
          type: integer
          format: int32
          readOnly: true
          description: Server-assigned revision number. Clients must not set this.
...
    Exercise:
      type: object
      description: |
        The blueprint/definition of an exercise (e.g. "Bench Press"), not a
        performed instance. Bundled/default exercises are server-owned;
        custom exercises are owned by the user who created them.
      required:
        - id
        - name
        - primaryMuscles
        - secondaryMuscles
        - instructions
        - tips
        - difficulty
        - equipmentNeeded
        - category
        - exerciseType
        - trackingType
        - source
        - deprecated
        - updatedAt
        - revision
      properties:
        id:
          type: string
          format: uuid
          description: Client-generated (CUSTOM) or server-generated (DEFAULT) UUIDv7.
        name:
          type: string
          description: Name of the exercise, e.g. "Bench Press", "Barbell Squat".
        primaryMuscles:
          type: array
          description: Primary muscle groups targeted by this exercise. Can include multiple, e.g. for compound movements.
          items: { $ref: '#/components/schemas/MuscleGroup' }
        secondaryMuscles:
          type: array
          description: Secondary muscle groups engaged to assist, but not the main focus.
          items: { $ref: '#/components/schemas/MuscleGroup' }
        description:
          type: [string, "null"]
          description: Detailed description of how to perform the exercise - form cues, breathing tips, general guidance.
        instructions:
          type: array
          description: Step-by-step instructions for performing the exercise, one string per step.
          items: { type: string }
        tips:
          type: array
          description: Key tips for performing the exercise correctly - helpful hints and safety advice.
          items: { type: string }
        difficulty:
          $ref: '#/components/schemas/DifficultyLevel'
        equipmentNeeded:
          type: array
          description: Equipment required to perform this exercise. Can include multiple items, e.g. barbell + bench.
          items: { $ref: '#/components/schemas/Equipment' }
        category:
          $ref: '#/components/schemas/ExerciseCategory'
        exerciseType:
          $ref: '#/components/schemas/ExerciseType'
        trackingType:
          $ref: '#/components/schemas/ExerciseTrackingType'
        videoUrl:
          type: [string, "null"]
          format: uri
          description: URL to an external video demonstration (e.g. YouTube).
        thumbnailUrl:
          type: [string, "null"]
          format: uri
          description: URL to a thumbnail image of the exercise, for preview in lists and cards.
        source:
          allOf:
            - $ref: '#/components/schemas/EntitySource'
          description: Whether this is a default/bundled (or server-provided) exercise, or one the user created.
        ownerId:
          type: [string, "null"]
          description: Firebase UID of the creator, for CUSTOM exercises. Null for DEFAULT rows.
        derivedFromId:
          type: [string, "null"]
          format: uuid
          description: |
            Id of the DEFAULT exercise this one was duplicated from, for a
            future "duplicate and edit" action on a default entry. Not used
            yet.
        deprecated:
          type: boolean
          description: |
            Whether this exercise has been superseded and should no longer
            appear in lists, search, or filters. Existing references to it
            remain resolvable.
        updatedAt:
          type: integer
          format: int64
          readOnly: true
          description: Server-assigned last-change time, epoch milliseconds. Clients must not set this.
        deletedAt:
          type: [integer, "null"]
          format: int64
          description: |
            Tombstone time, epoch milliseconds; non-null means this row
            is deleted. On push, this is the client's delete signal, not
            a timestamp the server trusts: set it to any non-null value
            to delete the row, leave it null otherwise. The server
            always stamps its own current time on write and ignores the
            value sent - what comes back from pull is the authoritative
            deletedAt. Deleting a DEFAULT exercise this way is rejected
            the same as any other write to one (FORBIDDEN) - only CUSTOM
            rows can be deleted.
        revision:
          type: integer
          format: int32
          readOnly: true
          description: Server-assigned revision number. Clients must not set this.
    WorkoutTemplate:
      type: object
      description: |
        A reusable blueprint of planned exercises (sets/reps/rest) that a
        workout session can be started from. Unlike WorkoutSession, carries
        no actual performance data. Syncs as a whole aggregate together with
        its template exercises.
      required:
        - id
        - title
        - templateExercises
        - source
        - deprecated
        - updatedAt
        - revision
      properties:
        id:
          type: string
          format: uuid
          description: Client-generated (CUSTOM) or server-generated (DEFAULT) UUIDv7.
        title:
          type: string
          description: User-defined title, e.g. "Push Day", "Leg Day", "Full Body".
        templateExercises:
          type: array
          description: |
            Exercises planned in this template with their target sets/reps.
            Order reflects the planned sequence of exercises.
          items: { $ref: '#/components/schemas/TemplateExercise' }
        source:
          allOf:
            - $ref: '#/components/schemas/EntitySource'
          description: |
            Whether this is a default/bundled (or server-provided) template,
            or one the user created. Default templates cannot be edited or
            deleted, only hidden.
        ownerId:
          type: [string, "null"]
          description: Firebase UID of the creator, for CUSTOM templates. Null for DEFAULT rows.
        derivedFromId:
          type: [string, "null"]
          format: uuid
          description: |
            Id of the DEFAULT template this one was duplicated from, for a
            future "duplicate and edit" action on a default entry. Not used
            yet.
        deprecated:
          type: boolean
          description: |
            Whether this template has been superseded and should no longer
            appear in lists, search, or filters. Existing references to it
            remain resolvable.
        updatedAt:
          type: integer
          format: int64
          readOnly: true
          description: Server-assigned last-change time, epoch milliseconds. Clients must not set this.
        deletedAt:
          type: [integer, "null"]
          format: int64
          description: |
            Tombstone time, epoch milliseconds; non-null means this
            template is deleted. On push, this is the client's delete
            signal, not a timestamp the server trusts: set it to any
            non-null value to delete the template, leave it null
            otherwise. The server always stamps its own current time on
            write and ignores the value sent - what comes back from pull
            is the authoritative deletedAt. Deleting a DEFAULT template
            this way is rejected the same as any other write to one
            (FORBIDDEN) - only CUSTOM templates can be deleted.
        revision:
          type: integer
          format: int32
          readOnly: true
          description: Server-assigned revision number. Clients must not set this.
...
    ExerciseCategory:
      type: string
      description: General category of an exercise, classifying it by training modality.
      enum:
        - STRENGTH
        - CARDIO
        - FLEXIBILITY
        - PLYOMETRIC
        - OLYMPIC_LIFTING
        - STRONGMAN
        - CALISTHENICS
    MuscleGroup:
      type: string
      description: Major muscle group targeted by an exercise, used for primary/secondary muscle classification.
      enum:
        - CHEST
        - BACK
        - LEGS
        - SHOULDERS
        - ARMS
        - CORE
        - OTHER
...
    DifficultyLevel:
      type: string
      description: Difficulty level of an exercise, to help users pick exercises suited to their experience.
      enum:
        - BEGINNER
        - INTERMEDIATE
        - ADVANCED
...
  responses:
    BadRequest:
      description: Malformed request or failed validation. error is VALIDATION_FAILED.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthenticated:
      description: Missing, expired, or invalid token. error is UNAUTHENTICATED.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    CursorExpired:
      description: |
        Cursor predates the tombstone retention window (90 days): a
        deletion the client hasn't seen yet has been pruned from this
        account's retained history and can no longer be replayed. error is
        CURSOR_EXPIRED. The client must discard all local user data and
        perform a full resync (call pull again with no cursor).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests. error is RATE_LIMITED. Respect Retry-After before retrying.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalError:
      description: Unexpected server failure. error is INTERNAL. Include traceId when reporting.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }